requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Technical Analysis
TA-Lib==0.4.28

# Testing
hypothesis==6.98.0
pytest==7.4.4
pytest-asyncio==0.23.3

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# Cache
redis==5.0.1
hiredis==2.3.2

# Async Tasks
celery==5.3.6
celery[redis]==5.3.6

# AI/ML
transformers==4.36.2
torch==2.1.2
google-generativeai==0.3.2

# Data Collection
python-binance==1.0.19
requests==2.31.0
aiohttp==3.9.1
beautifulsoup4==4.12.3
feedparser==6.0.10
tweepy==4.14.0
praw==7.7.1
pytrends==4.9.2

# Security & Authentication
cryptography==42.0.0
pyjwt==2.8.0
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
email-validator==2.1.0

# Reporting
reportlab==4.0.9
weasyprint==60.2

# CORS
python-cors==1.0.0

